moved main file to seperate folder, implemented db init
